buypop up err

+++ dashboard/App.js
@@ -1368,7 +1368,7 @@ async function execSwap() {
 
 
     const slippageBps = Math.floor(slippagePct * 100);
-    const deadline = Math.floor(Date.now() / 1000) + 1200;
+    const deadline = BigInt(Math.floor(Date.now() / 1000) + 1200);
     const to = connectedAddress;
 
     showLoading("Preparing swap…");
